requirements.txt
atproto
python-dotenv
feedparser
google-genai
